requirements.txt
aiohttp==3.9.5
pyyaml==6.0.1
rich==13.7.1
aiofiles==23.2.1
certifi==2024.8.30
cloudscraper==1.2.71
selenium==4.15.0
undetected-chromedriver==3.5.5